+++ api-contract/Abi/toV1.ts
@@ -1,11 +1,11 @@
 // Copyright 2017-2022 @polkadot/api-contract authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { ContractMetadataV0, ContractMetadataV1 } from 'https://deno.land/x/polkadot@0.2.3/types/interfaces/index.ts';
-import type { Registry } from 'https://deno.land/x/polkadot@0.2.3/types/types/index.ts';
+import type { ContractMetadataV0, ContractMetadataV1 } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
+import type { Registry } from 'https://deno.land/x/polkadot/types/types/index.ts';
 
-import { convertSiV0toV1 } from 'https://deno.land/x/polkadot@0.2.3/types/mod.ts';
-import { objectSpread } from 'https://deno.land/x/polkadot@0.2.3/util/mod.ts';
+import { convertSiV0toV1 } from 'https://deno.land/x/polkadot/types/mod.ts';
+import { objectSpread } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 interface Named {
   name: unknown;
@@ -21,6 +21,10 @@ function v0ToV1Names (all: Named[]): unknown[] {
 }
 
 export function v0ToV1 (registry: Registry, v0: ContractMetadataV0): ContractMetadataV1 {
+  if (!v0.metadataVersion.length) {
+    throw new Error('Invalid format for V0 (detected) contract metadata');
+  }
+
   return registry.createType('ContractMetadataV1', objectSpread({}, v0, {
     spec: objectSpread({}, v0.spec, {
       constructors: v0ToV1Names(v0.spec.constructors),