+++ api-contract/Abi/toV4.ts
@@ -1,0 +1,9 @@
+// Copyright 2017-2022 @polkadot/api-contract authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import type { ContractMetadataV3, ContractMetadataV4 } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
+import type { Registry } from 'https://deno.land/x/polkadot/types/types/index.ts';
+
+export function v3ToV4 (registry: Registry, v3: ContractMetadataV3): ContractMetadataV4 {
+  return v3;
+}