+++ api-contract/base/types.ts
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Observable } from 'https://esm.sh/rxjs@7.5.6';
-import type { SubmittableExtrinsic } from 'https://deno.land/x/polkadot@0.2.3/api/submittable/types.ts';
-import type { ApiTypes, ObsInnerType } from 'https://deno.land/x/polkadot@0.2.3/api/types/index.ts';
-import type { AccountId } from 'https://deno.land/x/polkadot@0.2.3/types/interfaces/index.ts';
+import type { SubmittableExtrinsic } from 'https://deno.land/x/polkadot/api/submittable/types.ts';
+import type { ApiTypes, ObsInnerType } from 'https://deno.land/x/polkadot/api/types/index.ts';
+import type { AccountId } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
 import type { AbiMessage, BlueprintOptions, ContractCallOutcome, ContractOptions } from '../types.ts';
 
 export interface MessageMeta {