+++ types/interfaces/bridges/types.ts
@@ -1,12 +1,12 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable */
 
-import type { BitVec, Bytes, Enum, Null, Struct, U8aFixed, Vec, bool, u32, u64 } from 'https://deno.land/x/polkadot@0.2.3/types-codec/mod.ts';
-import type { ITuple } from 'https://deno.land/x/polkadot@0.2.3/types-codec/types/index.ts';
-import type { BlockHash } from 'https://deno.land/x/polkadot@0.2.3/types/interfaces/chain/index.ts';
-import type { MultiSignature } from 'https://deno.land/x/polkadot@0.2.3/types/interfaces/extrinsics/index.ts';
-import type { AuthorityList, SetId } from 'https://deno.land/x/polkadot@0.2.3/types/interfaces/grandpa/index.ts';
-import type { AccountId, Balance, BlockNumber, H256, Header, MultiSigner, Weight } from 'https://deno.land/x/polkadot@0.2.3/types/interfaces/runtime/index.ts';
+import type { BitVec, Bytes, Enum, Null, Struct, U8aFixed, Vec, bool, u32, u64 } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
+import type { ITuple } from 'https://deno.land/x/polkadot/types-codec/types/index.ts';
+import type { BlockHash } from 'https://deno.land/x/polkadot/types/interfaces/chain/index.ts';
+import type { MultiSignature } from 'https://deno.land/x/polkadot/types/interfaces/extrinsics/index.ts';
+import type { AuthorityList, SetId } from 'https://deno.land/x/polkadot/types/interfaces/grandpa/index.ts';
+import type { AccountId, Balance, BlockNumber, H256, Header, MultiSigner, Weight } from 'https://deno.land/x/polkadot/types/interfaces/runtime/index.ts';
 
 /** @name BridgedBlockHash */
 export interface BridgedBlockHash extends H256 {}