+++ types/interfaces/contractsAbi/definitions.ts
@@ -3,7 +3,7 @@
 
 import type { Definitions } from '../../types/index.ts';
 
-import { objectSpread } from 'https://deno.land/x/polkadot@0.2.3/util/mod.ts';
+import { objectSpread } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 // order important in structs... :)
 /* eslint-disable sort-keys */
@@ -106,6 +106,7 @@ const spec = {
     events: 'Vec<ContractEventSpecV2>',
     docs: 'Vec<Text>'
   },
+  ContractContractSpecV4: 'ContractContractSpecV3',
   ContractDisplayName: 'SiPath',
   ContractEventParamSpecV0: {
     name: 'Text',
@@ -177,6 +178,7 @@ const spec = {
 };
 
 const ContractMetadataV0 = {
+  metadataVersion: 'Text',
   types: 'Vec<Si0Type>',
   spec: 'ContractContractSpecV0'
 };
@@ -196,6 +198,8 @@ const ContractMetadataV3 = {
   spec: 'ContractContractSpecV3'
 };
 
+const ContractMetadataV4 = ContractMetadataV3;
+
 const ContractProjectInfo = {
   source: 'ContractProjectSource',
   contract: 'ContractProjectContract'
@@ -207,7 +211,7 @@ const latest = {
   ContractEventParamSpecLatest: 'ContractEventParamSpecV2',
   ContractMessageParamSpecLatest: 'ContractMessageParamSpecV2',
   ContractMessageSpecLatest: 'ContractMessageSpecV2',
-  ContractMetadataLatest: 'ContractMetadataV3'
+  ContractMetadataLatest: 'ContractMetadataV4'
 };
 
 export default {
@@ -218,12 +222,14 @@ export default {
     ContractMetadataV1,
     ContractMetadataV2,
     ContractMetadataV3,
+    ContractMetadataV4,
     ContractMetadata: {
       _enum: {
         V0: 'ContractMetadataV0',
         V1: 'ContractMetadataV1',
         V2: 'ContractMetadataV2',
-        V3: 'ContractMetadataV3'
+        V3: 'ContractMetadataV3',
+        V4: 'ContractMetadataV4'
       }
     },
     ContractProjectV0: objectSpread({ metadataVersion: 'Text' }, ContractProjectInfo, ContractMetadataV0),